Move LimitQueryDeterminismAnalyzer.Analysis to top level
